requirements.txt
ascend-io-sdk==0.2.36
ipykernel==6.12.1
pyspark==3.2.1